Added a make target for creating a single header
On certains systems, managing multiple headers is hard, e.g. Windows.
having a single include with everything makes things easier got folks
using the lib on those platforms (e.g. me:))

A new directory with a makefile and make target has been created for
generating a single include. This make target concatinates the contents
of the include folder (though based on what's included in
include/gethin.hpp). Another gethin.hpp with all contents is created in
the single_include directory by this make target. This is also used by
the tests and the example.

A already 'compiled' single_include is also included in the repo for
folks not having access to sed, make etc.